+++ oauthlib/openid/connect/core/grant_types/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 """
-oauthlib.oauth2.rfc6749.grant_types
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+oauthlib.openid.connect.core.grant_types
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 """
 from __future__ import unicode_literals, absolute_import
 
@@ -10,7 +10,7 @@ from .implicit import ImplicitGrant
 from .base import GrantTypeBase
 from .hybrid import HybridGrant
 from .exceptions import OIDCNoPrompt
-from oauthlib.openid.connect.core.grant_types.dispatchers import (
+from .dispatchers import (
     AuthorizationCodeGrantDispatcher,
     ImplicitTokenGrantDispatcher,
     AuthorizationTokenGrantDispatcher